+++ docs/changelog.md
@@ -1,5 +1,8 @@
 # TBD
 
+### Changes
+* Updated the implemented methods for KnownKeysOnly<ITestSuiteServiceServer> to callback google_protobuf_empty_pb.Empty when needed
+
 # 0.4.2
 
 ### Changes